[AArch64] Split gcc.target/aarch64/sve/reduc_strict_3.c
This patch splits gcc.target/aarch64/sve/reduc_strict_3.c into one
test per function, so that it's easier to see what each scan is
matching and also so that we no longer rely on the number of times
that each dump message is printed.

The patch also generalises the tests to work with scalable vectors.
I think the test probably predates support for variable-length
loop-aware SLP.

2019-10-31  Richard Sandiford  <richard.sandiford@arm.com>

gcc/testsuite/
	* gcc.target/aarch64/sve/reduc_strict_3.c: Split all but the
	first function out into...
	* gcc.target/aarch64/sve/reduc_strict_4.c,
	* gcc.target/aarch64/sve/reduc_strict_5.c,
	* gcc.target/aarch64/sve/reduc_strict_6.c,
	* gcc.target/aarch64/sve/reduc_strict_7.c,
	* gcc.target/aarch64/sve/reduc_strict_8.c,
	* gcc.target/aarch64/sve/reduc_strict_9.c: ...these new tests.
	Test for scalable vectors instead of 256-bit vectors.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@277681 138bc75d-0d04-0410-961f-82ee72b054a4